Calls Sanity API directly to fetch the state list.
During build, in order to support SSG with dynamic routes, we need to fetch the list of states from Sanity.
We need to do this directly from the Sanity API instead of the app router api because the routes do not exist until after the build process is completed.
This commit updates the stateService to call the Sanity API directly to fetch the list of states and SSG render the pages.