Removes client side rendering from lender/agent cards.
State was only be used to manage toggling the state of the read more/less button. This was replaced with a css only solution that uses a hidden checkbox to track the state of the read more/less button. The checkbox is hidden and the label is styled to look like a button. When the label is clicked, the checkbox is checked and the text is expanded. When the label is clicked again, the checkbox is unchecked and the text is collapsed.